requirements.txt
fastapi[all]
sqlalchemy
pydantic_settings
asyncpg
python-jose
bcrypt==4.0.1
passlib[bcrypt]
pyjwt
alembic
Pillow
loguru
apscheduler